move maildetail to containers and rename the filter reducer
