+++ src/components/App.js
@@ -4,7 +4,7 @@ import {
   BrowserRouter as Router, Route, Switch,
 } from 'react-router-dom';
 import RecipesList from '../containers/recipesList';
-import MealDetails from './MealDetails';
+import MealDetails from '../containers/MealDetails';
 
 function App() {
   return (
